[CS-B-055] Объединила классы и поменяла нумерацию сценариев
