--- server/auth/src/test/java/ru/cfuv/cfuvscheduling/auth/bdd/then/UserRegistrationThen.java
@@ -1,28 +1,0 @@
-package ru.cfuv.cfuvscheduling.auth.bdd.then;
-
-import io.cucumber.java.en.Then;
-import ru.cfuv.cfuvscheduling.auth.bdd.AuthTestContext;
-import ru.cfuv.cfuvscheduling.auth.bom.AccountResponse;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-public class UserRegistrationThen {
-
-    private final AuthTestContext testContext = AuthTestContext.getINSTANCE();
-
-    @Then("request registration should be processed successfully")
-    public void requestShouldBeProcessedSuccessfully() {
-        Object responseObject = testContext.getResponseObject();
-        Class responseClass = testContext.getResponseClass();
-
-        assertNotNull(responseObject);
-        assertEquals(responseClass, AccountResponse.class);
-    }
-
-    @Then("request registration won't be processed due to {}")
-    public void requestWontBeProcessedDueTo(String exceptionClass) {
-        Class exceptionResponseClass = testContext.getExceptionClass();
-        assertEquals(exceptionClass, exceptionResponseClass.getSimpleName());
-    }
-}
