+++ server/auth/src/test/java/ru/cfuv/cfuvscheduling/auth/bdd/when/UserAuthorizationWhen.java
@@ -36,4 +36,27 @@ public class UserAuthorizationWhen {
     public void serverGotAuthorizationRequestWithEmptyUsernameAndPassword(String password) {
         serverGotAuthorizationRequestWithUsernameAndPassword(null, password);
     }
+
+    @When("server got registration request with username {} and password {}")
+    public void serverGotRegistrationRequestWithUsernameAndPassword(String username, String password) {
+        try {
+            AccountForm accountForm = new AccountForm();
+            accountForm.setUsername(username);
+            accountForm.setPassword(password);
+            AccountResponse response = authController.registerUser(accountForm);
+            testContext.setResponseObject(response, AccountResponse.class);
+        } catch (Exception e) {
+            testContext.setExceptionObject(e, e.getClass());
+        }
+    }
+
+    @When("server got registration request with empty username and password {}")
+    public void serverGotRegistrationRequestWithEmptyUsernameAndPassword(String password) {
+        serverGotRegistrationRequestWithUsernameAndPassword(null, password);
+    }
+
+    @When("server got registration request with username {} and empty password")
+    public void serverGotRegistrationRequestWithUsernameAndEmptyPassword(String username) {
+        serverGotRegistrationRequestWithUsernameAndPassword(username, null);
+    }
 }
